add cmd to expose openshift service, add condition to remove namespace only if it's not monitoring

+++ scripts/cluster-helpers.sh
@@ -186,9 +186,11 @@ function minikube_terminate() {
 	rm ${HPO_DEPLOY_MANIFEST}
 	echo
 
-	echo
-	echo "Removing HPO namespace"
-	kubectl delete ns ${hpo_ns}
+	if [ ${hpo_ns} != "monitoring" ]; then
+		echo
+		echo "Removing HPO namespace"
+		kubectl delete ns ${hpo_ns}
+	fi
 }
 
 ###############################  utilities  #################################
