+++ scripts/openshift-helpers.sh
@@ -43,6 +43,8 @@ function openshift_deploy() {
 		# Indicate deploy failed on error
 		exit 1
 	fi
+	# Expose HPO service so that it can be accessed outside the cluster
+	oc expose svc/hpo -n $hpo_ns
 
 	# Get the HPO application port in openshift
 	OPENSHIFT_IP=$(${kubectl_cmd} get pods -l=app=hpo -o wide -n ${hpo_ns} -o=custom-columns=NODE:.spec.nodeName --no-headers)
@@ -85,7 +87,9 @@ function openshift_terminate() {
 	rm ${HPO_DEPLOY_MANIFEST}
 	echo
 
-	echo
-	echo "Removing HPO namespace"
-	kubectl delete ns ${hpo_ns}
+	if [ ${hpo_ns} != "monitoring" ]; then
+		echo
+		echo "Removing HPO namespace"
+		kubectl delete ns ${hpo_ns}
+	fi
 }
